requirements.txt
flask
markupsafe
mysqlclient
pymysql
cryptography
flask-sqlalchemy
requests
oauthlib
pyOpenSSL
Flask-Login
gunicorn
psycopg2
flask-injector